% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.linreg}
\alias{coef.linreg}
\title{coef function shows regression coefficients in linreg}
\usage{
\method{coef}{linreg}(data, ...)
}
\arguments{
\item{data}{is a data frame}
}
\description{
coef function shows regression coefficients in linreg
}
