% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coefA}
\alias{coefA}
\title{coef function shows regression coefficients in linreg}
\usage{
coefA(data, ...)
}
\arguments{
\item{data}{is a data frame}
}
\description{
coef function shows regression coefficients in linreg
}
