% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{Advance r lab 4 using rc class related to linear regression}
\arguments{
\item{formula}{is a formula given as an input}

\item{data}{is a data frame}
}
\value{
class
}
\description{
Advance r lab 4 using rc class related to linear regression
}
\section{Fields}{

\describe{
\item{\code{regressionsCoefficients}}{a linear regression coefficient of linreg}

\item{\code{fittedValues}}{a fitted value of linreg}

\item{\code{residuals}}{a residual of linreg}

\item{\code{degreesOfFreedom}}{a degree of freedom of linreg}

\item{\code{residualVariance}}{a residual variance of linreg}

\item{\code{varianceRegressionCoefficients}}{a variance Regression Coefficients of linreg}

\item{\code{tValuesForEachCoefficient}}{a t Values For Each Coefficient}

\item{\code{pValues}}{a p Values of linreg}

\item{\code{call}}{is a call of function name}
}}


