% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.linreg}
\alias{plot.linreg}
\title{plot function shows graphical reperentation}
\usage{
\method{plot}{linreg}(data, ...)
}
\arguments{
\item{data}{is a adata frame}
}
\description{
plot function shows graphical reperentation
}
