% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{pred.linreg}
\alias{pred.linreg}
\title{pred function shows fitted values in linreg}
\usage{
pred.linreg(data, ...)
}
\arguments{
\item{data}{is a data frame}
}
\description{
pred function shows fitted values in linreg
}
