% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{predA}
\alias{predA}
\title{pred function shows fitted values in linreg}
\usage{
predA(data)
}
\arguments{
\item{data}{is a data frame}
}
\value{
thios function returns fitted value
}
\description{
pred function shows fitted values in linreg
}
