% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.linreg}
\alias{print.linreg}
\title{print function shows linear regression coefficients}
\usage{
\method{print}{linreg}(data, ...)
}
\arguments{
\item{data}{is a data frame}
}
\description{
print function shows linear regression coefficients
}
