% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{printA}
\alias{printA}
\title{print function shows linear regression coefficients}
\usage{
printA(data, ...)
}
\arguments{
\item{data}{is a data frame}
}
\description{
print function shows linear regression coefficients
}
