% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid.R
\name{residA}
\alias{residA}
\title{resid function shows residual in linreg}
\usage{
residA(data, ...)
}
\arguments{
\item{data}{is a data frame}
}
\value{
this function returns residual
}
\description{
resid function shows residual in linreg
}
