% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summery.R
\name{summary.linreg}
\alias{summary.linreg}
\title{summary function shows summary of linreg}
\usage{
\method{summary}{linreg}(data, ...)
}
\arguments{
\item{data}{is a data frame}
}
\description{
summary function shows summary of linreg
}
